Move residence_time to Transitions; it needs no jump data
residence_time() only reads transitions.states and sites.labels, so
Jumps is the wrong owner. Moving it to Transitions keeps the method
co-located with the data it depends on and removes any implication that
it is filtered by minimal_residence. Test moved to transitions_test.py
using vasp_transitions directly.
